requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
torchmetrics>=1.2.0
albumentations>=1.3.0
einops>=0.7.0
timm>=0.9.0
lpips>=0.1.4
